remove duplicate test logic
